Attempting to fix Activity launching issues
* Attempting to fix setting com.onesignal.NotificationOpened.DEFAULT
to DISABLE does not prevent app from resuming when app is backgrounded.
   - issue #1366
* Left notes on attempts made and results found on different scenarios
and different versions of Android. Mostly issues on Android 4.4 and 5.0
* The hope was a fix for these issue was going to be straightforward and
small but it is looking like we will need forking code for different
versions of Android to fix with the current approach.
* Instead of continuing to fix the current approach the path forward is
to use a "Reverse Activity Trampolining" approach. Some of the learnings
found in this commit however might be useful to implementing new
approach but that change will be in another PR.