Add synchronization to  to fix possible race condition
Fixes a null pointer exception occurring due to attempt to call `showView` on a null `messageView` object, likely due to it being set to null by a background thread resulting in a race condition. By locking the `messageView` to one thread at a time, we can prevent this issue from occurring.