Fix padding on BGR IMG notifications for RTL
Correct the title and body 4dp padding from "left" to "start"
to support both RTL and LTR languages. This panding is important so
there is a bit of room between the large icon or the edge of the screen
depending on if the text is LTR or RTL, in that order .

Changed `layout_marginStart` to `left` since the SDK does not support
a different image based on if the device is RTL or LTR. Along with this
change the documenation should instruct developers to always put
their app icon on the left side.