proguard -assumenosideeffects compatibility
Fix compatibility issue with apps that use -assumenosideeffects with
java.lang.Class.getName() with progurad or dexguard. No issue with R8.

Addressed this by adding a new private hasClass method and utilized it
in all existing class exist helper methods.
This method has the "Keep" annotation on it so proguard will never
change it's code and will not make a wrong assumptions about side
effects.