switch sdk to use Gradle's rich versions
This allows us to define a required version range while providing a
preferred tested exact version.
This gives the consumer consistent versions when building that
won't change if a newer version of a library like firebase-messaging is
released.

We want to keep version ranges to be as compatible as possible with
other dependencies declared for the exact module we use. This means if
an older version of a library OneSignal depends on is also required by
the app or its dependencies Gradle will accommodate within the ranges
that overlap.

This also fixes build errors when the app uses
`strict-version-matcher-plugin` or uses
the `google-services` Gradle plugins without setting
`disableVersionCheck = true`. This is because the Google plugins require
a version scheme the provides consistent builds.

I have tested end-to-end to confirm the facts above are correct.