Use `currentTime` instead of `currentThreadTime`
When we check `isNotificationWithinTTL` to determine if we should display it, we should use `currentTimeMillis` (milliseconds since epoch) instead of `currentThreadTimeMillis` to determine the current time that will be used in the logic. The google sent_time is also in milliseconds since epoch.

During some testing, `currentThreadTimeMillis` is equivalent to almost 0 seconds due to it being the time a thread has been active.

While working on this, I looked at other places that used thread time inaccurately and changed to using `currentTimeMillis` instead.