Fix group compatibility Android 6 and lower
After testing Android 6 in the Android Studio emulator it was found
that using setGroupAlertBehavior with GROUP_ALERT_CHILDREN breaks
the heads-up and sound for notifications that are grouped. Since the
GROUP_ALERT_CHILDREN flag is required for newer versions of Android
to fix a different bug with heads up notifications from showing old
notifications in the group we needed to add an Android version runtime
check to select the correct flag.

We could report this bug to Google and they might be able to fix it
in a future AndroidX release however due to the age of Android 6 and
it's limitations it is unlikely it will be fixed. So this workaround
will mostly likely need to stay in place until we drop support for
Android 6.

See the last commit for more details on why GROUP_ALERT_CHILDREN is
needed for Android 8+. This also addresses the failing test form the
last commit.