rm all the reverse bits from Activity Trampolining
This commit focuses on removing all logic that setups up multiple
Activities and flags to build a "Reverse Activity Trampoline".

This is Part 1 of required changes, this commit results in the app not
opening, resuming, or opening launch URLs, the behavior of the click
being tracked is still working via the invisible
NotificationOpenedReceiver Activity. This gives us a clean starting
state for the next commit where we we will add logic to the notification
open event to Activity Trampoline forward to the intended Activity.