Don't retry get params on thread interruption
If the app process is shutting down we don't want to keep retrying the
remote params network request. No real world problem have been
discovered but it's possible it could be create some unexpected states
while the app process is being shutdown.

In the context of our tests, this also fixes an issue where this
retrying does not stop when we mock fail this network request. This
oddly enough didn't seem to be an issue for the one pre-existing test
using failGetParams but became a consistent test carry over issue when a
2nd test used this, which was added in the last commit.