rm waiting on params network call on notifi open
This fixes a bug where the app would not be brought to the foreground
when tapping on a notification if the app could not connect to
onesignal.com or was offline during the full life time of the app
process.
This also fixes another a compatibility issue with Xiaomi devices where
the app would not foreground if the app process was dead and also no
task exists in the recent list.

It is critical we don't wait for a network call for the notification
open logic this method does the work to bring the app to the foreground.
This get params network call isn't needed before we handle the
notification open logic anyway.

This network wait was present in 4.0.0 through 4.4.x but was not an
issue in 4.5.0 when a "Reverse Activity Trampoline" was setup. However
in PR #1581 we are switching back to a standard Activity Trampoline and
we don't want to reintroduce this bug.

A failing test was added for this in a previous commit, which now
passes.