rm FLAG_ACTIVITY_NO_HISTORY from open URL
Removing FLAG_ACTIVITY_NO_HISTORY so the user can see the Activity to go
back to it if they background the app.

Also removed other flags so when the user press the back button on the
device they are taken to the last Activity in the app if the URL is a
deeplink. If the browser is open then it is it's own task and pressing
the back button will then take them to the home screen, however the
app's task will remind in the list so the user can always get back to
where they left off in the app.

There isn't any details into why these flags were added, they date back
to the 2.0.0 release of this SDK:
https://github.com/OneSignal/OneSignal-Android-SDK/commit/ce5ad7d07e9d34df207e7514dd3fff98fb26f184#diff-2708f86f77e539cab647ecede9d5fcbead3bf0f954d7f9e536eafbc78ba760d6R792
However given that the user experience is much better in these cases
this is a better default. If the URL is a deeplink the app developer
could always add other Activity setting to their AndroidManifest.xml for
additional control.