Corrected reporting of FCM exceptions
Since we moved to FCM's newer Task API we did not correctly unwrap the
nested root causes to parse the root error message. Correct this along
with some refactoring to make this easier to read and to support another
error type.

Also accounted for AUTHENTICATION_FAILED which we didn't handle before.
This can happen if you boot an Android Studio emulator for the first
time right after creating it and run your app. We should retry as it
only takes a short amount of time for the emulator to initialize with
Google's servers.