Adding the release drafter action to create draft release notes
As PRs are merged to main release drafter will create a draft release and fill out the release notes template.
PRs will be bucketed into fix, feature, or improvement based on labels. We need to tag our PRs with the appropriate labels to get the automatic bucketing. However PRs can always be moved around when we create the release PR.