add public CallbackThreadManager class
This new public API allows changing the default thread used to fire aync
events, observers, and callbacks for the rest of the OneSignal public
API.

This is intended for advanced use-case, motivated
mostly a need of the OneSignal Unity wrapper SDK. Specifically to
avoid cases where Unity may cause the main UI thread to wait on a
background thread when calling back into Unity. Given how the binding
between C# and Java works here, there is no way for the C# code to
switch threads itself, so this API has to be available to do so.