Don't hydrate device info for a push subscription model
When hydrating an existing push subscription model, use existing local device-scoped information instead of remote information for:
- sdk
- deviceOS
- carrier
- appVersion
This information should always come from the local device. The reason for this change is that on a new session, we may detect a change to one of these properties and we do not want to overwrite it with old remote data from the get_user response.