Add device properties to Subscription Model
Subscription Model now has sdk, deviceOS, carrier, and appVersion as properties so that they can persist.
The getter is defaulted to "" (the empty string) as these properties do not exist prior to v5.0.5. They are only read when detecting changes so it is ok that they default to "" as the empty string will not be sent.