fix issue with @RobolectricTest failing
The first class with @RobolectricTest would run fine, but then every
test class after in the run would fail with a looper already started
error.

This fix was discovered by looking through Robolectric's
RobolectricTestRunner and SandboxTestRunner clases. Most of the
pre-existing kotest-extensions-robolectric code seemed to be reripped
from those 2 classes and I notice this was something this extension was
not doing. Looper's are linked to a thread so this seems related,
however I don't understand any more than this to explain why this fixes
the error.