use com.tdunning:json which is closer to AOSP
org.json:json is a different implementation than Android AOSP, swapped it
out with com.tdunning:json which seems to be identical to AOSP. This is
important as we want the test to behave as close as possible to a real
Android device.

Added some tests to ensure if we ever change JSON libraries again
we don't run into hard to debug problems later.