prototype of adding input params to getService
Adding input params to getService, this is particularly useful for the
injecting Context into ApplicationService so it doesn't have to handle
it being null.

However passing input parameters to getService would only be a good idea
if it returns a new instance every time or if it uses the parameter to
give a instance based on it. Nether of these is what we want for
ApplicationService however.

Instead of this pattern I think it would be better to allow creating
instances outside of the ServiceProvide and allow adding them after
the initial ServiceProvider has been built. Or if we don't build it
until we have a Context, and we have a way to our own instances at
build time.

Context is also maybe something we don't want in the
dependency injection system anyway? The idea is that most classes
shouldn't refer to Context directly, instead a class should be built to
abstract away the Android API. However in the end there is a spot
where the the rubber meets the road and that will need Context injected
at some point, but if a class accepts Context as a constructor it
probably should be the only one.