rm refresh User on login of same external_id
We already refresh when the app is cold started, so this results
in two GET calls back-to-back for apps that call login every time
the app is opened the by the user. We are still refreshing the user if
switch user happens.

Didn't add any unit test since there is no existing test infrastructure
to test code in OneSignalImp. Also such code should be refactored out
into a different class to make it more testable anyway.