add backoff to operations when retrying
If we got a 5xx response from the OneSignal backend we were retrying
immediately. This result wasn't positive for the end-user, device,
or OneSignal's backend.

To ensure this works as expected we updated the retrying test to
confirm the delay is called.

Also to ensure this commit works as intended I ran this on an Android 14
emu and ensured it waits the right amount of time when there is 500
response. Also after we stop getting a 500 error the queue successfully
resumes and all pending tags are even added too.