call initWithContext on all entry points
We call OneSignal.initWithContext on most entry points but some were
missed. Consistency is important as differences can lead to edge case
bugs that are hard to reproduce and debug.

One recent example of this was with ReceiveReceiptWorker, its
inconsistency lead to a hard to debug issue that was uncovered by
commit 893381b05fbd55eef74dbfc663dd27cceb098901.

In this commit an issue was discovered in NotificationGenerationWorker.
It almost always runs after receiving a push, but not guaranteed. It
was checking for OneSignal.isInitialized but instead should always call
initWithContext as it can be an entry point to the app process.

The ADMMessageHandlerJob may or may not have had an issue, but we
matched up the initWithContext behavior to be safe.