switch to kotest-extensions-android
Move off our custom patched kotest-extensions-robolectric to the
kotest-extensions-android library. This is a newly released library
based off the patches we shared with the original developer. Switching
over to kotest-extensions-android means we can remove this complexity
from our project and can easily update the library to get fixes.

We had to bump our minSdkVersion to 21 (Android 5.0) to use this library
however we only officially supported Android 5.0 and newer since v5 of
our SDK anyway.