make taskQueueWaitingForInit thread safe
There are two calls to startPendingTasks that can happen on two threads,
which means sometimes taskQueueWaitingForInit.poll() can return null and
throws a NPE when it is given to submit. To handle this we are making
taskQueueWaitingForInit thread safe by calling synchronized on. Doing so
handles this case but also rare cases where something is being added as
we process the queue.