fix multiple enqueues skipping waitForWake
We renamed waitForWake to waitForNewOperationAndExecutionInterval to
describe that it is doing two things.

Fixed the logic where enqueuing two or more operations would skip our
executionInterval logic. Added test coverage for this logic, as well
as a 2nd test to ensure flush can still skip waiting for it.

Lastly cleaned up the force state in the main processQueueForever loop,
we were able to fully encapsulate it into
waitForNewOperationAndExecutionInterval. The other part of the logic is
once we start processing the queue we want to do so until it's empty
again, which ops != null is the only checked needed in
processQueueForever to cover that.