only fetch IAMs if app is in the foreground
We want to prevent wasted resources polling for IAMs durning times where
it is unlikely they will be seen. Two changes are being made; 1. ensure
we only poll when the isInForeground is true. 2. poll the first time
the app is foregrounded, incase it is not when the SDK is initialized.