fix onSubscriptionChanged not firing for local-
This is a common case where updates (such as optedIn) should
still propagate even if we haven't sent the POST /users create
call yet. Motivation for this test was a bug was discovered
where calling OneSignal.User.pushSubscription.optIn() was not
prompting for notification permission if it was called before
the create User network call finished.