attempt at simplifying OperationsRepo grouping
DO NOT MERGE, prototyping idea only

In this commit attempted to simplify the OperationRepo by not allowing
out of order execution of operations. Simplifying the grouping down to
one key as well.

While this change can simply things it does so by lowering efficiency,
as not as many operations are grouped together. It also doesn't get us
to more ideal model, of having a list of User each other a list of
operations. The changes attempted here are on the medium side
and comes with risk, so opting for a smaller fix for now, with a follow
up later to do the larger improvement.