fix missing modify keys on some operations
All operations must have at least a create or modify key, this is
required so operations are not grouped together when they should not be.

This fixes a bug where calling addTag() back-to-back with login() would
cause the login to be dropped. This is because
SetTagOperation.modifyComparisonKey was "" which matched
LoginUserOperation.modifyComparisonKey as it was also "" and both these
operations we returned by OperationRepo.getGroupableOperations. Both
operations were given to UpdateUserOperationExecutor which it would skip
the LoginUserOperation silently.

This mistake originated from f1d204e3ebaff51aa8f4a7239bc681e918d75078
which wasn't in the 5.0.0-beta4 release, but was introduced in 5.0.0.
The intent of PR #1794 is still preserved in this commit, we just simply
fixed the bug noted above.