recover from a stalled user in the OperationRepo
Commit 7fe2010dd1326b1a1e47619c0f3ee17b919aa221 fixed a bug with login
operations being drop in some scenarios, however it doesn't fix those
already in the bad state.

This comment get us out of the bad state by generating a replacement
LoginUserOperation. See the comments in the code for more details.

This comment was manually tested, ensuring it gets us out of the bad
state, and updates works afterwords.