don't re-create user on failed del alias attempts
We incorrectly assumed that getting a 404 or a 410 from
DELETE users/by/$aliasLabel/$aliasValue/identity/$aliasLabelToDelete
means the user is gone. It could mean either the User and/or the Alias
is gone, either way the end state is the same, the alias no longer exists
on that User. If the User was really delete we will know on the next
User operation, and that will handle the create User.

This fixes a bug where the OperationRepo gets stuck in a loop trying to
create the user over-and-over again if OneSignal.User.deleteAlias is
called twice in a row with the same label.

With the current SDK logic, the retrying was be done with a backoff,
so retrying won't have had much of an impact on the OneSignal's
backend load on 5.1.7 and later, but did on 5.0.0 to 5.1.6.