create test to prove operOp's enqueue doesn't wait
This commit simply adds a test to prove the following issue with
the current implementation of OperationRepo.

Once processQueueForever starts processing operations it will execute
them back-to-back until it can't process any more. This is done
intently as the idea is sync all changes to the backend quickly, after
waiting a set amount of time for batching. However nothing is in place
to account for something continually adding new operations to the repo
while it's in this mode. Some misbehaving app could be adding
operations in a tight loop as fast or faster than the queue could
exclude them.