save network calls on OpRepo by adding bucketing
Buckets
Purpose: Bucketing is a pattern we are using to help save network
calls. It works together with opRepoExecutionInterval to define
a time window operations can be added to the bucket.

When enqueue() is called it creates a new OperationQueueItem with it's
bucket = enqueueIntoBucket. Just before we start processing a bucket we
enqueueIntoBucket++, this ensures anything new that comes in while
executing doesn't cause it to skip the opRepoExecutionInterval delay.

NOTE: Bucketing only effects the starting operation we grab.
The reason is we still want getGroupableOperations() to find
other operations it can execute in one go (same network call).
It's more efficient overall, as it lowers the total number of
network calls.

This address the failing test "operations enqueued while repo is
executing should be executed only after the next
opRepoExecutionInterval" we added in commit
0988977c07b1af558242b9359f92a15500e6ee33