never set OperationRepo.enqueue's flush to true
OneSignal.login() was setting OperationRepo.enqueue to true, while this
means the OneSignal backend is updated up to 5 seconds faster it means
a possible increase in load, as not having a delay means we may not
have batched as many network requests.

There is a high chance we will re-introduce the flush via a new public
commit() method, so it wasn't remove in this commit. But before doing
so we need budget rules so it can't be abused.

One oddly I notice here is the login had a flush, but the logout did
not. This seems like a bug, but either way we don't want to flush
anywhere now.