prevent adding executionInterval & postCreateDelay
After we wait for the opRepoPostCreateDelay time call to waiter.wake()
would cause waitForNewOperationAndExecutionInterval() to wake up as
we expect, however it was then waiting for the full time on
opRepoExecutionInterval as well.
To solve this we simply subtract the time we already waited from
opRepoExecutionInterval. With the current values this means we
fully skip opRepoExecutionInterval so we only wait
opRepoPostCreateDelay.