restart wait time on every OperationRepo.enqueue()
Every time something new is enqueued we will not restart waiting for
batches timer. The motivation is to prevent a misbehaving app from
continuously making network calls.
We are basically saying wait for the "the dust to settle" or
"the water is calm" to ensure the app is done making updates.

FUTURE: Highly recommend not removing this "the dust to settle" logic,
as it ensures any app stuck in a loop can't cause continuous network
requests. If the delay is too long for legitimate use-cases then allow
tweaking the opRepoExecutionInterval value or allow commitNow() with a
budget.