use the retryAfterSeconds in OperationRepo
Use the retryAfterSeconds to stall processing in the OperationRepo. We
wait in the OperationRepo as this allows more operations to be grouped
together, as we getGroupableOperations will be called after the delay.

There are number of other places that use HTTPClient other than the
OperationRepo, we will address those in a follow up commit.