delay OneSignal network requests if Retry-After
When we get a Retry-After we delay all future requests until after this
time as passed. To be safe we assume the server's Retry-After value
means all traffic should be delayed until then. If we wanted finer
grain control we could work with the the backend to decide which
endpoints should be affected.