add HTTP header OneSignal-Install-Id
OneSignal-Install-Id is a UUIDv4 locally generated on the device and
added to as an HTTP Header to all calls made to OneSignal's backend.
This allows the OneSignal's backend know where traffic is coming
from, no matter if the SubscriptionId or OneSignalId changes or isn't
available yet.

State for the new installId is encapsulated in a new InstallIdService
class, where it handles generating the id and persisting it. Tests
were also added to ensure the persisting behavior works as expected.