ensure internalEnqueue can't add the same operation
This fixes a bug where loadSavedOperations would duplicate operations
in OperationRepo.queue. This is because it is possible for operations
to be added before it is started and these are also persisted to disk.

Improved OperationRepoTests by making mockOperationModelStore modify
a in memory list. This was required to write a test to ensure
loadSavedOperations wasn't duplicating operations.