fix loadSavedOperations index out of bounds issue
Since things can be added to the queue before loadSavedOperations is
called it has to account for duplicate entries. It did do this, however
it was missing the logic to account for not advancing the index on
duplicates so an out of bounds was possible.

We solved the problem by only incrementing the index if it wasn't a
duplicate, however this implementation has a future landmine. If
something ever removes something from the queue that
loadSavedOperations added and it is still executing index problems can
still happen. This scenario never happens now, but a new feature to
OperationRepo or a refactor could introduce the problem. A fast follow
is recommend so this landmine isn't left here.