swap EventProducer for CompletableDeferred
This change fixes a rare bug where RecoverFromDroppedLoginBug may not
run. While very rare this could happen if start was called on
OperationRepo first and it finished before
RecoverFromDroppedLoginBug.start was called.

We swapped out EventProducer for a CompletableDeferred to address this
issue. CompletableDeferred is a better fit as we simply need to wait
until something is initialized, and it will never fire more than once.

We also refactored RecoverFromDroppedLoginBugTests to account for the
change, but improved the testing to ensure the operation itself is
correct.