simply loadSavedOperations logic
Reversing the list means we insert each item at index 0 to keep the
order correct. Always inserting at index 0 means we no longer need to
keep an index state, which both simples the code but also means we
don't depend on state which may be wrong sometimes, noted in the
warning comment that was removed in this commit. This reversed()
logic for list used in a few places in this file already to solve this
problem.