Xiaomi fix notification click not foregrounding
This fixes a bug where tapping on a OneSignal notification on a Xiaomi
device doesn't bring the app to the foreground.
Xiaomi has very strict activity trampolining rules on notification
open. We have to start the next activity before the current activity is
finished.

Tested on Xiaomi Redmi 6A device with MIUI 11.0.8 (Android 9) as well
as an Android 14 emulator.