fix minification compatibility with stub classes
If the app project had android.enableR8.fullMode=true
(default true with AGP 8+) the stub classes that start with
"misconfigured" would removed. This would cause a
"could not be instantiated" exception to be thrown when initializing
as the code assumes there would be some class that implements the
interface, an example INotificationsManager.

The stub classes are removed in the minification step since they are
created with reflection via ServiceProvider.getService(). Anytime
reflection is used there you must add special Proguard consumer rules
to keep those classes. In this case we only need to make sure the class
itself exists, so we use "<init>(...);" in the rule which means keep
the default constructor.