rm permission usages in AndroidSupportV4Compat
Switched to using Jetpack's implementation. The original purpose of the
this class was not require Android Support Library 23, but we are long
past needing to support that.

Null signature is more strict with Jetpack, so some of the code needed
a small refactor.

There is one last function, getColor, which didn't belong there that
will be cleaned up in a follow up commit.