Update `InAppBackendService` to handle `retryLimit`
Motivation: if the offset is not null, we are telling the server give us the IAMs while ensuring read-your-write consistency.

If it is present, we will retry the request based on two response headers:
1. `Retry-After`
2. `Retry-Limit`

If we hit the limit, we will make a final request *without* the offset, in order to tell the server it gives up, just give me the best-effort IAMs.