fix ending an already ended session
# What
Fixes an issue where invalid REST API requests were made to both
outcomes/measure and User updates with session_time were bing made in
the background. This is only present when location sharing is enabled
in the SDK (in addition to the app and end-user allowing it) and the
end-user opens the app at any point after receiving a notification.
This resulted in 400 errors in the outcomes/measure case and no-op User
updates that wasting resources, the later doesn't depend on receiving
notifications. The outcomes/measure is also additive as the bad
requests are cached and retried on app open.

# How
Add a state check to SessionService to prevent ending an already ended
session.

Also address a previous session never ending correctly when the app is
cold started.

Lastly, added comments pointing out the caveats of IBackgroundService
to help prevent future wrong assumptions about when backgroundRun() is
run.

# Follow up
Since there are cached invalid outcomes/measure requests a migration
needs to be created in the SDK to remove them for apps who were
affected by the bug.