`IamFetchReadyCondition` logic nit & better comments
Motivation: add details explaining logic.

We should also resolve the condition immediately if one of the tokens is equal to the `RYW_TOKEN_NOT_PROVIDED` const.

We DON'T want to block IAMs if we turn off the feature or a regression occurs that excludes the token from the endpoint response body.