fix grouping when setColor is used in NSE
This applies the color of the most recent notification added to the
group to the summary notification. Without this logic the grouped
notifications not only display the wrong color (the default is used)
but also Android decides to display the small icon on the left of each
notification as well. It seems the AOSP tries to be "smart" and if the
icons are the same it just omits them on children.

The logic is applied to both the SDK groupless / auto grouping logic
as well as ones defined by push or overridden in the NSE.