fix rare app not opening from notification bug
In rare cases when tapping on a notification the app would not open,
as well as other issues like the click not being counted either.

The issue was due to a race condition between
NotificationListener.start and processing the click event. To fix
and to avoid this mistake in the future moved all the code from
NotificationListener into NotificationLifecycleService. There
wasn't a clear set of responsibility difference between these
to classes so this is also a win to remove complexity.

Was able to reproduce the issue 1/5 times on an Android 14
(no extension level) emulator. After testing 20 times I am
no longer able to reproduce the issue now.