fix concurrency with modifying log listeners
Avoid ConcurrentModificationException when someone adds or removes a
log listener while the stack is still inside the onLogEvent method.

Addressed by switching from an ArrayList to CopyOnWriteArrayList.
This is a good fit as the listener list will rarely be modified, but
will have a ton of reads.

Added test to prove and ensure this issue is now avoided.