fix crash when adding alias with some labels
When passing some labels to User.addAlias, such as "AL" this could
cause a crash when starting the app the next time. This is due to
casting the wrong type as initializeFromJson's code used a contains
instead of an equals which can result in finding the wrong function.

To correct this switched this to equals("get$property", true) instead
and also added a check for it.parameterCount == 0 as we only want to
cast based on getters.

Also added tests for IdentityModel with a number of labels to ensure
this code changes covers common label values.

Lastly also removed the if (method == null) check to simplify the code,
as the when statement can handle this case.