Switch publishing to the Central Portal
OSSRH is shutting down soon so we need to migrate to the Central Portal
for publishing.
* There are no changes to those consuming this SDK, only for mainliners
to publish.

Add the "plugin com.vanniktech.maven.publish" plugin to publish to
the Central Portal.

To publish the following command must be used now:
./gradlew publishToMavenCentral --no-configuration-cache \
-PmavenCentralUsername=USER_TOKEN_USERNAME_HERE \
-PmavenCentralPassword=USER_TOKEN_PASSWORD_HERE \
-Psigning.keyId=ID_HERE \
-Psigning.password=PASSWORD_HERE \
-Psigning.secretKeyRingFile=/FULL/PATH/TO/GPG_FILE_HERE

Can switch to publishAndReleaseToMavenCentral once we do a release
manually and test end-to-end.