rm Amazon in-app purchase tracking
Amazon doesn't have an official API for SDKs to track IAP, as a result
we have been using reflection to hook into internal API's. This is
extremely fragile, it breaks on every patch version of Amazon's IAP
library.

Due to the fact the market is moving away IAP through app stores and
Amazon's pullback on bring the Amazon store to other platforms
(like Windows) the importance has shrunk so low the maintenance isn't
worth it anymore.