DO NOT MERGE
Tabling the following for now but keeping in-case we need to go
down this road later.  Attempted to use
@AutoService(AndroidInstrumentation::class) automatically to setup
a custom crash handler for the OneSignal SDK, however
ServiceLoader.load(AndroidInstrumentation::class.java) would not
pick up our class.
I found suggestions that the could be due to META-INF/services/
not containing the class in your build / aar, however after trying a
few options to fix the issue none worked. This approach does fit the
pattern of the other opentelemetry-android instrumentations however,
their approach is more complex and based on the issue I am having with
ServiceLoader.load not working seems more fragile to silently breaking
then I would like.