fix: remove throwing on init timeout
Starting in 5.4.0 calls such as OneSignal.Notification or
OneSignal.Location would throw if they waited on the SDK into to long.
This comment changes this to wait forever until init is complete, and
log instead if the time is taking longer than expected.
If these calls were done from the main thread an ANR is the better of
two evils and the app can recover, where an uncaught throw it can not.

This commit will bring the behavior the similar behavior in 5.1.x.
The only difference is you will now never see an ANR with
initWithContext, however if you call other parts of the SDK from the
main thread, such as OneSignal.Notification, you will continue to see
ANRs, just at a different stacktrace. To avoid these ANRs completely
call all other SDK methods outside of the main thread.