fix: ensure OperationRepo is executed in order
Made a new internal coroutine scope with a single thread to the
OperationRepo to ensure operations are alway executed in the order they
are queued in.

This mostly fixes the flaky OperationRepoTest class.

Release 5.4.0 may have introduced bugs related to this, however they
would have been rare and very hard to reproduce.